requirements.txt
streamlit
chromadb
sentence-transformers
pypdf
regex
requests
rank_bm25
pyyaml
python-docx
